Add useValidateCampaignWithCountryCodes hook.
